requirements-delete-dataset.txt
# Requirements for BigQuery dataset deletion script
google-cloud-bigquery>=3.15.0
python-dotenv>=1.0.0
google-auth>=2.23.0
